chore: Get rid of `serde_qs` entirely
