+++ Cargo.toml
@@ -45,7 +45,6 @@ sea-orm-migration = { version = "1", default-features = false }
 secrecy = { version = "0.8", default-features = false }
 serde = { version = "1", default-features = false }
 serde_json = { version = "1", default-features = false }
-serde_qs = { version = "0.13", default-features = false }
 serde_with = { version = "3", default-features = false }
 service = { path = "crates/service" }
 strum = { version = "0.26", default-features = false }
