+++ crates/rest-api/src/forms/mod.rs
@@ -4,12 +4,10 @@
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 mod interval;
+pub mod multi_value_items;
 mod pagination;
-pub mod qs_query;
-mod strict_qs_query;
 mod timestamp;
 
 pub use interval::Interval;
 pub use pagination::Pagination;
-pub use qs_query::QsQuery;
 pub use timestamp::Timestamp;
